Group files into modules (#74)
* group files into modules

* fix import errors

* style: [CI] cleanup python notebooks

* style: [CI] format

* fix test

* fix ci
