+++ apperception/data_types/query_type.py
