+++ apperception/scenic_util.py
@@ -5,8 +5,9 @@ from typing import Any, Dict, Iterable, List, Tuple
 
 import numpy as np
 import pandas as pd
-from box import Box
 from pyquaternion import Quaternion
+
+from apperception.data_types import Box
 
 CREATE_ITEMTRAJ_SQL = """
 CREATE TABLE IF NOT EXISTS Item_General_Trajectory(
