+++ apperception/utils/__init__.py
@@ -1,5 +1,21 @@
 from . import F
+from .add_recognized_objects import add_recognized_objects
+from .compute_heading import compute_heading
+from .df_to_camera_config import df_to_camera_config
 from .fn_to_sql import fn_to_sql
 from .query_to_str import query_to_str
+from .recognize import recognize
+from .reformat_bbox_trajectories import reformat_bbox_trajectories
+from .scenic_generate_df import scenic_generate_df
 
-__all__ = ["F", "fn_to_sql", "query_to_str"]
+__all__ = [
+    "F",
+    "fn_to_sql",
+    "query_to_str",
+    "reformat_bbox_trajectories",
+    "add_recognized_objects",
+    "compute_heading",
+    "recognize",
+    "scenic_generate_df",
+    "df_to_camera_config",
+]
