+++ apperception/utils/compute_heading.py
